Fixed some linting problems.
