+++ php/example_code/aws_utilities/AWSServiceClass.php
@@ -39,5 +39,4 @@ abstract class AWSServiceClass
         }
         return $result;
     }
-
 }
