+++ php/example_code/dynamodb/dynamodb_basics/GettingStartedWithDynamoDB.php
@@ -23,7 +23,7 @@
  *
  **/
 
-# snippet-start:[php.example_code.ddb.dynamodb_basics.scenario]
+# snippet-start:[php.example_code.dynamodb.basics.scenario]
 namespace DynamoDb\Basics;
 
 require_once "vendor/autoload.php";
@@ -38,8 +38,8 @@ use GuzzleHttp\Client as GuzzleClient;
 use ZipArchive;
 use function AwsUtilities\testable_readline;
 
-class GettingStartedWithDynamoDB {
-    
+class GettingStartedWithDynamoDB
+{
     public function run()
     {
         echo("--------------------------------------\n");
@@ -66,7 +66,7 @@ class GettingStartedWithDynamoDB {
         $service->dynamoDbClient->waitUntil("TableExists", ['TableName' => $tableName]);
         echo "table $tableName found!\n";
 
-#snippet-start:[php.example_code.dynamodb.basics.putItem]
+# snippet-start:[php.example_code.dynamodb.basics.putItem]
         echo "What's the name of the last movie you watched?\n";
         while (empty($movieName)) {
             $movieName = testable_readline("Movie name: ");
@@ -88,7 +88,7 @@ class GettingStartedWithDynamoDB {
             ],
             'TableName' => $tableName,
         ]);
-#snippet-end:[php.example_code.dynamodb.basics.putItem]
+# snippet-end:[php.example_code.dynamodb.basics.putItem]
 
         echo "How would you rate the movie from 1-10?\n";
         $rating = 0;
@@ -135,7 +135,7 @@ class GettingStartedWithDynamoDB {
             $file = \GuzzleHttp\Psr7\Utils::tryFopen($movieZipName, 'w');
             $guzzle->request("get", $url, ['sink' => $file]);
 
-            $zip = new ZipArchive;
+            $zip = new ZipArchive();
             $extractPath = ".";
             if ($zip->open($movieZipName) !== true) {
                 echo "Could not open/find the zip file. Please check your file system permissions.";
@@ -151,8 +151,8 @@ class GettingStartedWithDynamoDB {
         $limit = 0;
         $service->writeBatch($tableName, $batch);
 
-#snippet-start:[php.example_code.dynamodb.basics.getItem]
-#snippet-start:[php.example_code.dynamodb.basics.key]
+# snippet-start:[php.example_code.dynamodb.basics.getItem]
+# snippet-start:[php.example_code.dynamodb.basics.key]
         $key = [
             'Key' => [
                 'title' => [
@@ -163,34 +163,34 @@ class GettingStartedWithDynamoDB {
                 ],
             ]
         ];
-#snippet-end:[php.example_code.dynamodb.basics.key]
+# snippet-end:[php.example_code.dynamodb.basics.key]
 
         $movie = $service->getItemByKey($tableName, $key);
         echo "\nThe movie {$movie['Item']['title']['S']} was released in {$movie['Item']['year']['N']}.\n";
-#snippet-end:[php.example_code.dynamodb.basics.getItem]
-#snippet-end:[php.example_code.dynamodb.basics.updateItem]
+# snippet-end:[php.example_code.dynamodb.basics.getItem]
+# snippet-start:[php.example_code.dynamodb.basics.updateItem]
         echo "What rating would you like to give {$movie['Item']['title']['S']}?\n";
         $rating = 0;
         while (!is_numeric($rating) || intval($rating) != $rating || $rating < 1 || $rating > 10) {
             $rating = testable_readline("Rating (1-10): ");
         }
         $service->updateItemAttributeByKey($tableName, $key, 'rating', 'N', $rating);
-#snippet-end:[php.example_code.dynamodb.basics.updateItem]
+# snippet-end:[php.example_code.dynamodb.basics.updateItem]
 
         $movie = $service->getItemByKey($tableName, $key);
         echo "Ok, you have rated {$movie['Item']['title']['S']} as a {$movie['Item']['rating']['N']}\n";
 
-#snippet-start:[php.example_code.dynamodb.basics.deleteItem]
+# snippet-start:[php.example_code.dynamodb.basics.deleteItem]
         $service->deleteItemByKey($tableName, $key);
         echo "But, bad news, this was a trap. That movie has now been deleted because of your rating...harsh.\n";
-#snippet-end:[php.example_code.dynamodb.basics.deleteItem]
+# snippet-end:[php.example_code.dynamodb.basics.deleteItem]
 
         echo "That's okay though. The book was better. Now, for something lighter, in what year were you born?\n";
         $birthYear = "not a number";
         while (!is_numeric($birthYear) || $birthYear >= date("Y")) {
             $birthYear = testable_readline("Birth year: ");
         }
-#snippet-start:[php.example_code.dynamodb.basics.query]
+# snippet-start:[php.example_code.dynamodb.basics.query]
         $birthKey = [
             'Key' => [
                 'year' => [
@@ -199,7 +199,7 @@ class GettingStartedWithDynamoDB {
             ],
         ];
         $result = $service->query($tableName, $birthKey);
-#snippet-end:[php.example_code.dynamodb.basics.query]
+# snippet-end:[php.example_code.dynamodb.basics.query]
         $marshal = new Marshaler();
         echo "Here are the movies in our collection released the year you were born:\n";
         $oops = "Oops! There were no movies released in that year (that we know of).\n";
@@ -210,7 +210,7 @@ class GettingStartedWithDynamoDB {
         }
         echo ($display) ?: $oops;
 
-#snippet-start:[php.example_code.dynamodb.basics.scan]
+# snippet-start:[php.example_code.dynamodb.basics.scan]
         $yearsKey = [
             'Key' => [
                 'year' => [
@@ -228,9 +228,10 @@ class GettingStartedWithDynamoDB {
             $movie = $marshal->unmarshalItem($movie);
             echo $movie['title'] . "\n";
         }
-#snippet-end:[php.example_code.dyanmodb.basics.scan]
+# snippet-end:[php.example_code.dynamodb.basics.scan]
 
         echo "\nCleaning up this demo by deleting table $tableName...\n";
         $service->deleteTable($tableName);
     }
 }
+# snippet-end:[php.example_code.dynamodb.basics.scenario]
